server: [breaks backcompat] Eliminate race involving server.Draining().
There was a race condition when something called <-server.Draining() before server.ListenAndServe.
ListenAndServe is what sets the return value of server.Draining(), and if you call Draining before
that channel is setup by ListenAndServe, Draining() returns a channel that isn't closed when
ListenAndServe starts draining.  (It's actually just "nil".)

The new design makes you explicitly initialize the channel at server startup time, so you won't run
into this problem.

(The initalization could be fixed with explicit synchronization, but that leads to mixing channels
and mutexes, which always leads to weird deadlocks.  So I changed the API instead.  Also
complicating issues is that the tests call ListenAndServe more than once.  This isn't supported, but
I wanted it to keep working for the convenience of the tests.)